requirements.txt
openai>=1.0.0,<2
requests
python-dotenv
rich
tqdm
pandas

